requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
mediapipe>=0.10.13
streamlit>=1.29.0
numpy==1.26.4
Pillow>=10.0.0
pygame>=2.5.0
protobuf==4.25.3
